Don't parallax for reduced motion
