+++ app/components/Footer.tsx
@@ -2,7 +2,10 @@ export default function Footer() {
     return (
         <footer>
             <span>
-                <a href="https://github.com/olillin/olillin.com" target="_blank">
+                <a
+                    href="https://github.com/olillin/olillin.com"
+                    target="_blank"
+                >
                     See the source code on GitHub
                 </a>
             </span>
