+++ app/layout.tsx
@@ -5,7 +5,7 @@ import '@fortawesome/fontawesome-svg-core/styles.css'
 config.autoAddCss = false
 
 import './style.css'
-import ParallaxBackground from './components/ParallaxBackground'
+import ParallaxBackgroundEffect from './components/ParallaxBackgroundEffect'
 
 export const metadata: Metadata = {
     title: 'Olillin',
@@ -19,9 +19,7 @@ export default function RootLayout({
     return (
         <html lang="en">
             <body>
-                <ParallaxBackground>
-                    {children}
-                </ParallaxBackground>
+                <ParallaxBackgroundEffect>{children}</ParallaxBackgroundEffect>
             </body>
         </html>
     )
